+++ assets/js/navigation.js
@@ -159,8 +159,8 @@ export function panTracks(direction = 'left', speed = 0.1) {
   }
   // drawTrack will correct the window eventually, but let us not go negative at least
   if (pos.start < 1) {
+    pos.end = distance - pos.start
     pos.start = 1
-    pos.end = distance
   }
   drawTrack({ chrom: pos.chrom, start: pos.start, end: pos.end, drawTitle: false, exclude: ['cytogenetic-ideogram'] })
 }