+++ CHANGELOG.md
@@ -8,6 +8,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Document track processing and loading
 - OAuth authentication
+### Changed
+- Hide balanced variants
 
 ## [2.1.1]
 ### Added