+++ CHANGELOG.md
@@ -8,7 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Document track processing and loading
 - OAuth authentication
-### Changed
+- Use sample id instead of display name for variant retrieval
 - Keyboard pan speed increased
 
 ## [2.1.1]