+++ CHANGELOG.md
@@ -8,8 +8,10 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Document track processing and loading
 - OAuth authentication
+### Changed
 - Use sample id instead of display name for variant retrieval
-- 
+- Hide balanced variants
+
 ## [2.1.1]
 ### Added
 ### Changed